Move stuff to helpers file
